Fix tests close #412
