+++ client/mobilizations/widgets/__plugins__/pressure/components/index.js
@@ -1,4 +1,4 @@
-export { default } from './__pressure__.connected'
+export { default } from './__pressure__'
 export { default as PressureCount } from './pressure-count'
 export { default as PressureForm } from './pressure-form'
 export { default as PressureTellAFriend } from './pressure-tell-a-friend'
