+++ test/client/unit/mobilizations/widgets/components/__widget__.spec.js
@@ -37,7 +37,7 @@ describe('client/mobilizations/widgets/components/__widget__', () => {
 
   it('should render Form widget component if widget kind it is "form"', () => {
     wrapper.setProps(changeKind({ props, kind: 'form' }))
-    expect(wrapper.find('Connect(Form)')).to.have.length(1)
+    expect(wrapper.find('Form')).to.have.length(1)
   })
 
   it('should render Draft widget component if widget kind it is "draft"', () => {
